+++ georiviere/portal/tests/test_serializers/test_portal.py
@@ -37,23 +37,27 @@ class PortalSerializerTest(TestCase):
 
     def test_portal_content(self):
         data = self.serializer_portal.data
-        self.assertSetEqual(set(data.keys()), {'id', 'map', 'name', 'flatpages'})
+        self.assertSetEqual(set(data.keys()), {'id', 'map', 'name', 'flatpages', 'title', 'description', 'extent',
+                                               'max_zoom', 'min_zoom', 'main_color'})
         self.assertEqual(len(data['map']['group'][0]['layers']), 5)
 
     def test_portal_all_layers_grouped_content(self):
         data = self.serializer_portal_layers_all_group.data
         self.assertEqual(len(data['map']['group']), 1)
         self.assertEqual(data['map']['group'][0]['label'], 'Bar')
-        self.assertSetEqual(set(data.keys()), {'id', 'map', 'name', 'flatpages'})
+        self.assertSetEqual(set(data.keys()), {'id', 'map', 'name', 'flatpages', 'title', 'description', 'extent',
+                                               'max_zoom', 'min_zoom', 'main_color'})
         self.assertEqual(len(data['map']['group'][0]['layers']), 5)
 
     def test_portal_without_se_content(self):
         data = self.serializer_portal_without_spatial_extent.data
-        self.assertSetEqual(set(data.keys()), {'id', 'map', 'name', 'flatpages'})
+        self.assertSetEqual(set(data.keys()), {'id', 'map', 'name', 'flatpages', 'title', 'description', 'extent',
+                                               'max_zoom', 'min_zoom', 'main_color'})
         self.assertAlmostEqual(data['map']['bounds'][0], -5.5006, delta=4)
         self.assertAlmostEqual(data['map']['bounds'][3], 51.314, delta=4)
 
     def test_portal_with_flatpages_content(self):
         data = self.serializer_portal_with_flatpages.data
-        self.assertSetEqual(set(data.keys()), {'id', 'map', 'name', 'flatpages'})
+        self.assertSetEqual(set(data.keys()), {'id', 'map', 'name', 'flatpages', 'title', 'description', 'extent',
+                                               'max_zoom', 'min_zoom', 'main_color'})
         self.assertEqual(set(data['flatpages'][0]), {'order', 'url', 'title', 'hidden'})